Ajustes nos geradores de partículas e nos Sprites compostos
1) Geradores de partículas são derivados direto da CPIGObjeto e as partículas são criadas coo cópias dessa instância
2) geradores de partículas têm métodos específicos para definir limites das partículas
3) Pontos das posições dos sprites são do tipo double para permitir posicinamentos corretos a partir de rotação do objeto.
4) Sprites compostos

+++ include/CPIGItemComponente.h
@@ -213,7 +213,7 @@ public:
         return 0;
     }
 
-    void Move(int nx, int ny)override{
+    void Move(double nx, double ny)override{
         int dx = nx-pos.x;
         int dy = ny-pos.y;
         CPIGComponente::Desloca(dx,dy);
